prima_api/app/requirements.txt
async-timeout==5.0.1
celery==5.4.0
fastapi[standard]==0.112.2
flower==2.0.1
gunicorn==23.0.0
mysql-connector-python==9.1.0
pre-commit==3.8.0
python-dotenv==1.0.1
pymongo==4.8.0
redis==5.0.8
uvicorn==0.30.6


